backend/app/services/card_generator.py
import google.generativeai as genai
from app.core.config import get_settings
from typing import Optional

settings = get_settings()
genai.configure(api_key=settings.GOOGLE_API_KEY)

class CardGenerator:
    def __init__(self):
        self.model = genai.GenerativeModel('gemini-pro')

    async def generate_card(
        self,
        word: str,
        base_language: str,
        target_language: str,
        is_cloze: bool = False
    ) -> dict:
        # Generate translations and example sentence
        prompt = f"""
        Generate a language learning card for the word "{word}" from {base_language} to {target_language}.
        Provide:
        1. The translation of the word
        2. A natural example sentence using the word
        3. A literal translation of the example sentence
        4. A semantic translation of the example sentence

        Format the response as JSON with these fields:
        - targetWord: the translation
        - exampleSentence: the example sentence
        - literalTranslation: word-for-word translation
        - semanticTranslation: natural translation
        """

        response = await self.model.generate_content(prompt)
        card_data = response.text

        # TODO: Parse the response and handle errors
        # For now, return a mock response
        return {
            "word": word,
            "targetWord": "translated word",
            "exampleSentence": "Example sentence in target language",
            "literalTranslation": "Word for word translation",
            "semanticTranslation": "Natural translation",
            "audioUrl": None,  # TODO: Implement audio generation
            "imageUrl": None,  # TODO: Implement image generation
        }

    async def generate_audio(self, text: str, language: str) -> Optional[str]:
        # TODO: Implement audio generation using Gemini
        return None

    async def generate_image(self, word: str) -> Optional[str]:
        # TODO: Implement image generation using Gemini
        return None 